// enums/messages.ts
enum Messages {
    INVALID_INPUT = 'Invalid input data',
    SERVER_ERROR = 'Internal server error',
    UNAUTHORIZED = 'Unauthorized access',
    FORBIDDEN = 'Forbidden access',
    RESOURCE_NOT_FOUND = 'resource not found',
    RESOURCE_CREATED = 'resource created successfully',
    RESOURCE_UPDATED = 'resource updated successfully',
    RESOURCE_DELETED = 'resource deleted successfully',
    EMAIL_OR_PASSWORD_INVALID = 'Invalid email or password',
    INVALID_API_KEY = 'Invalid API key',
    ACCESS_DENIED = 'Access denied'
};

export { Messages };